+++ packages/ai-chat/src/chat/components-legacy/responseTypes/_imports.scss
@@ -16,4 +16,3 @@
 @use "iframe/imports" as responseType-iframe;
 @use "image/imports" as responseType-image;
 @use "options/imports" as responseType-options;
-@use "util/imports" as responseType-util;
