+++ packages/ai-chat/src/chat/components-legacy/responseTypes/card/CardItemComponent.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -13,7 +13,7 @@ import React from "react";
 
 import { HasRequestFocus } from "../../../../types/utilities/HasRequestFocus";
 import { LocalMessageItem } from "../../../../types/messaging/LocalMessageItem";
-import { BodyWithFooterComponent } from "../../BodyWithFooterComponent";
+import { BodyWithFooterComponent } from "../../../components/BodyWithFooterComponent";
 import {
   CardItem,
   MessageResponse,
