+++ packages/ai-chat/src/chat/components-legacy/responseTypes/util/_imports.scss
@@ -5,8 +5,6 @@
  *  LICENSE file in the root directory of this source tree.
  */
 
-@use "BodyMessageComponents";
 @use "ClickableImage";
-@use "FooterButtonComponents";
 @use "SearchResultBody";
 @use "citations/imports" as citations;
