+++ packages/ai-chat/src/chat/components/BodyMessageComponents.tsx
@@ -1,5 +1,5 @@
 /*
- *  Copyright IBM Corp. 2025
+ *  Copyright IBM Corp. 2025, 2026
  *
  *  This source code is licensed under the Apache-2.0 license found in the
  *  LICENSE file in the root directory of this source tree.
@@ -9,11 +9,11 @@
 
 import cx from "classnames";
 import React from "react";
-import { useSelector } from "../../../hooks/useSelector";
+import { useSelector } from "../hooks/useSelector";
 
-import { AppState } from "../../../../types/state/AppState";
-import { MessageTypeComponentProps } from "../../../../types/messaging/MessageTypeComponentProps";
-import { MessageResponseTypes } from "../../../../types/messaging/Messages";
+import { AppState } from "../../types/state/AppState";
+import { MessageTypeComponentProps } from "../../types/messaging/MessageTypeComponentProps";
+import { MessageResponseTypes } from "../../types/messaging/Messages";
 
 interface BodyMessageComponentsProps extends MessageTypeComponentProps {
   renderMessageComponent: (
