+++ packages/ai-chat/src/chat/components/_imports.scss
@@ -10,4 +10,6 @@
 @use "modals/imports" as modals;
 @use "util/imports" as utils;
 @use "homeScreen/imports" as homescreen;
+@use "BodyMessageComponents";
+@use "FooterButtonComponents";
 @use "util/TextBlock/TextBlock.scss" as textblock;
